Rewrite dial-request to always index by nonce
The nonce of a probe is essentially like a primary key that is
decided by the client. Thus, any event emitted by the handler
should be indexed by the provided nonce.

We can achieve this by using a `FuturesMap` instead of a `FuturesSet`.
This gives us access to the nonce even in the case that the actual
protocol times out.

With the nonce in place, we had to re-model the event returned to
the behaviour. Most importantly. we need to separate the different
kinds of errors:

- Complete execution of the protocol but address is not reachable
- Protocol was aborted mid-way
- Server does not support the protocol

We can't really do anything if the protocols is aborted so we just
represent this case with an `io::Error` that gets logged further up.

As a result, this means we can remove the `Option` from the event
emitted to the user and _always_ give them a `Multiaddr`.